requirements.txt
boto3
psycopg2-binary
apache-airflow-providers-amazon